Print path and env changes at the end
In order to prevent the same paths and env vars from being printed
as both restored and added, we now print the actual changes at the end.

This requires some bookkeeping to keep track of which ones were added
and which ones were removed.

Also fixes ReversePaths to return a new slice, instead of modifying the
current one.